Stable COPD Analysis
This is the code used for final analysis of all the home patient data (total duration 6 hours only), excluding first 10 minutes and getting the values that are mean, SDNN, SD1, SD2, SE for heart rate and respiratory rate. Note after the mean all calculations use the R-R interval for HR and breath-breath (BB) interval for respiratory rate. This was repeated for each of the participants